# generated/src/main/java/org/openapitools/api/MemoryApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ErrorResponse;
import org.openapitools.model.MemoryDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-11T18:58:48.001078900+05:30[Asia/Calcutta]")
@Validated
@Tag(name = "Memory", description = "the Memory API")
public interface MemoryApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/memories : Add new memory
     *
     * @param memoryDTO  (required)
     * @return Memory added successfully (status code 200)
     *         or Not authorized to add memory (status code 401)
     *         or Bad request (status code 400)
     *         or Unacceptable memory (status code 406)
     */
    @Operation(
        operationId = "addMemory",
        summary = "Add new memory",
        tags = { "Memory" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Memory added successfully"),
            @ApiResponse(responseCode = "401", description = "Not authorized to add memory", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "406", description = "Unacceptable memory", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "jwtAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/memories",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> addMemory(
        @Parameter(name = "MemoryDTO", description = "", required = true) @Valid @RequestBody MemoryDTO memoryDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/memories/{memoryId} : Delete user memory
     *
     * @param memoryId Unique id of memory to delete (required)
     * @return memory deleted successfully (status code 200)
     *         or Memory not found (status code 404)
     *         or Unauthorized access to memory (status code 401)
     */
    @Operation(
        operationId = "deleteMemories",
        summary = "Delete user memory",
        tags = { "Memory" },
        responses = {
            @ApiResponse(responseCode = "200", description = "memory deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Memory not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized access to memory", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "jwtAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/memories/{memoryId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> deleteMemories(
        @Parameter(name = "memoryId", description = "Unique id of memory to delete", required = true, in = ParameterIn.PATH) @PathVariable("memoryId") String memoryId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/memories : Get all the memories
     *
     * @return List of all the memories of user (status code 200)
     */
    @Operation(
        operationId = "getMemories",
        summary = "Get all the memories",
        tags = { "Memory" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of all the memories of user", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = MemoryDTO.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "jwtAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/memories",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<MemoryDTO>> getMemories(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : \"mem_8f3a21c9\", \"title\" : \"First Day at New Job\", \"description\" : \"Started my new role as a backend developer and met the team.\", \"mood\" : \"happy\", \"tags\" : [ \"career\", \"milestone\" ] }, { \"id\" : \"mem_8f3a21c9\", \"title\" : \"First Day at New Job\", \"description\" : \"Started my new role as a backend developer and met the team.\", \"mood\" : \"happy\", \"tags\" : [ \"career\", \"milestone\" ] } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /api/memories/{memoryId} : Update memory
     *
     * @param memoryId Unique id of memory to update (required)
     * @param memoryDTO  (required)
     * @return Update successful (status code 200)
     *         or Memory not found (status code 404)
     */
    @Operation(
        operationId = "updateMemory",
        summary = "Update memory",
        tags = { "Memory" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Update successful"),
            @ApiResponse(responseCode = "404", description = "Memory not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "jwtAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/memories/{memoryId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> updateMemory(
        @Parameter(name = "memoryId", description = "Unique id of memory to update", required = true, in = ParameterIn.PATH) @PathVariable("memoryId") String memoryId,
        @Parameter(name = "MemoryDTO", description = "", required = true) @Valid @RequestBody MemoryDTO memoryDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}